+++ SmartWarehouseWeb/src/main/java/pl/infoshare/SmartWarehouseWebApplication.java
@@ -1,4 +1,4 @@
-package pl.infoshare.SmartWarehouseWeb;
+package pl.infoshare;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
