+++ SmartWarehouseWeb/src/main/java/pl/infoshare/controller/MainController.java
@@ -1,4 +1,4 @@
-package pl.infoshare.SmartWarehouseWeb.controller;
+package pl.infoshare.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class MainController {
 
-    @GetMapping("/test")
+    @GetMapping("/")
     public String main() {
         return "index";
     }
