--- SmartWarehouseWeb/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-server.port=8090
