+++ SmartWarehouseWeb/src/test/java/pl/infoshare/SmartWarehouseWebApplicationTests.java
@@ -1,4 +1,4 @@
-package pl.infoshare.SmartWarehouseWeb;
+package pl.infoshare;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
